server/requirements.txt
flask
flask-cors
tensorflow==2.16.1
numpy
Pillow
h5py
keras
opencv-python
gunicorn
werkzeug
